requirements.txt
fastapi
uvicorn
streamlit
gitpython
python-dotenv
langchain
langchain-core
langchain-groq
pytest
PyGithub